# docs/specs/02-api/openapi-v1.yaml
openapi: 3.1.0
info:
  title: Car Film Admin API Baseline
  version: 0.1.0
  description: |
    Baseline API contract frozen from current server implementation (`admin-console/server.py`).

    Notes:
    - Internal APIs accept either `X-Api-Token` or `Authorization: Bearer <INTERNAL_API_TOKEN>`.
    - Admin APIs use login token (`Authorization: Bearer <session_token>`).
servers:
  - url: http://127.0.0.1:8080
...
paths:
...
  /api/login:
    post:
      tags: [Auth]
      summary: Login to admin APIs
      operationId: postLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginSuccessResponse'
        '401':
          description: Invalid username/password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOkStyle'
  /api/me:
    get:
      tags: [Auth]
      summary: Current authenticated user
      operationId: getMe
      security:
        - adminBearer: []
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOkStyle'
...
components:
  securitySchemes:
    adminBearer:
      type: http
      scheme: bearer
      bearerFormat: token
      description: Session token returned by `/api/login`.
    internalBearer:
      type: http
      scheme: bearer
      bearerFormat: internal-token
      description: INTERNAL_API_TOKEN passed as Bearer token.
    internalApiToken:
      type: apiKey
      in: header
      name: X-Api-Token
      description: INTERNAL_API_TOKEN header.
...
  schemas:
    GenericObject:
      type: object
      additionalProperties: true
...
    UserSafe:
      type: object
      additionalProperties: true
      properties:
        username:
          type: string
        name:
          type: string
        role:
          type: string
        permissions:
          $ref: '#/components/schemas/GenericObject'
...
    LoginRequest:
      type: object
      additionalProperties: false
      properties:
        username:
          type: string
        password:
          type: string
      required: [username, password]
    LoginSuccessResponse:
      type: object
      additionalProperties: false
      properties:
        ok:
          type: boolean
        token:
          type: string
        user:
          $ref: '#/components/schemas/UserSafe'
      required: [ok, token, user]
    MeResponse:
      type: object
      additionalProperties: false
      properties:
        ok:
          type: boolean
        user:
          $ref: '#/components/schemas/UserSafe'
      required: [ok, user]
...
    ErrorOkStyle:
      type: object
      additionalProperties: true
      properties:
        ok:
          type: boolean
        message:
          type: string
...